requirements.txt
pyTelegramBotAPI==4.14.0
pandas==2.0.3
requests==2.31.0
openpyxl==3.1.2
python-dotenv==1.0.0
